remove warnnings

+++ TitlelessWindow.m
@@ -2,7 +2,7 @@
 
 
 @implementation TitlelessWindow
-- (id)initWithContentRect:(NSRect)contentRect styleMask:(unsigned int)aStyle backing:(NSBackingStoreType)bufferingType defer:(BOOL)flag {
+- (id)initWithContentRect:(NSRect)contentRect styleMask:(NSUInteger)aStyle backing:(NSBackingStoreType)bufferingType defer:(BOOL)flag {
 	NSWindow* result = [super initWithContentRect:contentRect styleMask:NSBorderlessWindowMask backing:bufferingType defer:flag];
     [result setLevel: NSStatusWindowLevel];
     [result setAlphaValue:0.8];
